--- apps/festival/festival/src/app/dashboard/title/import/import.component.ts
@@ -1,10 +1,0 @@
-
-import { Component, ChangeDetectionStrategy } from '@angular/core';
-
-@Component({
-  selector: 'festival-title-import',
-  templateUrl: './import.component.html',
-  styleUrls: ['./import.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class TitleImportComponent {}
