--- apps/financiers/financiers/src/app/dashboard/title/import/import.component.html
@@ -1,2 +1,0 @@
-
-<import-container importType="titles"></import-container>
