+++ libs/import/src/lib/import-spreadsheet/import-spreadsheet.component.ts
@@ -2,35 +2,38 @@
 import { Component, Output, EventEmitter, ChangeDetectionStrategy, ChangeDetectorRef, Optional, Input } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
 
-import { Intercom } from 'ng-intercom';
 
 import { getMimeType } from '@blockframes/utils/file-sanitizer';
 import { SheetTab, importSpreadsheet } from '@blockframes/utils/spreadsheet';
 
-import { sheetRanges, SpreadsheetImportEvent, SpreadsheetImportType } from '../utils';
+import { sheetRanges, SpreadsheetImportType } from '../utils';
 
 
 const allowedMimeTypes = ['application/vnd.ms-excel', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet', 'application/vnd.oasis.opendocument.spreadsheet', 'text/csv'];
 
 @Component({
-  selector: 'import-spreadsheet',
+  selector: 'import-spreadsheet[importType]',
   templateUrl: './import-spreadsheet.component.html',
   styleUrls: ['./import-spreadsheet.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ImportSpreadsheetComponent {
 
   /** The type of spreadsheet import we expect, can be `titles`, `organizations`, or `contracts` */
   @Input() importType: SpreadsheetImportType = 'titles';
 
-  @Output() importEvent = new EventEmitter<SpreadsheetImportEvent>();
+  @Output() importEvent = new EventEmitter<SheetTab>();
 
   public sheets: SheetTab[] = [];
+
+  public file: {
+    name: string;
+    size: number;
+  };
 
   constructor(
     private snackBar: MatSnackBar,
     private cdRef: ChangeDetectorRef,
-    @Optional() private intercom: Intercom,
   ) { }
 
   importSpreadsheet(files: FileList | File) {
@@ -57,23 +60,19 @@ export class ImportSpreadsheetComponent {
 
     const reader = new FileReader();
     reader.addEventListener('loadend', () => {
+      this.file = file;
       const buffer = new Uint8Array(reader.result as ArrayBuffer);
       this.sheets = importSpreadsheet(buffer, sheetRanges[this.importType]);
       this.cdRef.markForCheck();
-    })
+      this.importEvent.next(this.sheets[0]);
+    });
     reader.readAsArrayBuffer(file);
   }
 
-  next() {
-    // trigger the import event to tell parent component go to the next mat-stepper step
-    this.importEvent.next({ sheet: this.sheets[0], importType: this.importType });
-  }
-
-  removeFile() {
+  remove() {
+    console.log('remove');
+    delete this.file;
     this.sheets = [];
-  }
-
-  openIntercom() {
-    return this.intercom.show();
+    this.cdRef.markForCheck();
   }
 }
