+++ libs/import/src/lib/import-spreadsheet/import-spreadsheet.module.ts
@@ -9,9 +9,11 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatSelectModule } from '@angular/material/select';
 import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatChipsModule } from '@angular/material/chips';
 
 // Libraries
 import { ImageModule } from '@blockframes/media/image/directives/image.module';
+import { DownloadPipeModule } from '@blockframes/media/file/pipes/download.pipe';
 import { ViewExtractedElementsModule } from '../view-extracted-elements/view-extracted-elements.module';
 
 // Components
@@ -28,12 +30,14 @@ import { ImportSpreadsheetComponent } from './import-spreadsheet.component';
 
     // Material
     MatIconModule,
+    MatChipsModule,
     MatButtonModule,
     MatSelectModule,
     MatFormFieldModule,
 
     // Libraries
     ImageModule,
+    DownloadPipeModule,
     ViewExtractedElementsModule,
   ],
   exports: [ ImportSpreadsheetComponent ]
