--- libs/import/src/lib/import.component.html
@@ -1,13 +1,0 @@
-﻿
-<import-spreadsheet *ngIf="!fileType" [importType]="importType" (importEvent)="next($event)"></import-spreadsheet>
-
-<import-view-extracted-movies *ngIf="importType === 'titles'" #viewExtractedComponent></import-view-extracted-movies>
-<import-view-extracted-contracts *ngIf="importType === 'contracts'" #viewExtractedComponent></import-view-extracted-contracts>
-<import-view-extracted-organizations *ngIf="importType === 'organizations'" #viewExtractedComponent></import-view-extracted-organizations>
-
-<section *ngIf="fileType" fxLayout="column" fxLayoutAlign="center center">
-  <p>Please make sure everything is correct. Once you submit your titles they will instantaneously appear on the Archipel Market platform.</p>
-  <button test-id="cancel-import" mat-stroked-button color="warn" (click)="back()">
-    <mat-icon svgIcon="block"></mat-icon><span>Cancel</span>
-  </button>
-</section>
