--- libs/import/src/lib/import.component.scss
@@ -1,13 +1,0 @@
-.stepper-buttons {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-section {
-    margin-bottom: 20px;
-}
-
-p {
-    width: 50%;
-}
