--- libs/import/src/lib/view-extracted-elements/movies/view-extracted-movies.component.html
@@ -1,24 +1,0 @@
-<article>
-  <h1>Please make sure everything you want to import is correct before finalizing your submission.</h1>
-  <section class="help-section" fxLayout="column" fxLayoutAlign="center">
-    <p>Please save your titles first then review it and if you feel it's complete publish it to marketplace.</p>
-    <a [routerLink]="['./']" (click)="openIntercom()">Need help ?</a>
-  </section>
-  <ng-container *ngIf="moviesToUpdate.data.length || moviesToCreate.data.length else isLoading">
-    <section class="result-tables">
-      <ng-container *ngIf="moviesToCreate.data.length">
-        <h3>Titles to Create</h3>
-        <import-table-extracted-movies [rows]="moviesToCreate" mode="create"></import-table-extracted-movies>
-      </ng-container>
-      <ng-container *ngIf="moviesToUpdate.data.length">
-        <h3>Titles to Update & Submit</h3>
-        <import-table-extracted-movies [rows]="moviesToUpdate" mode="update"></import-table-extracted-movies>
-      </ng-container>
-    </section>
-  </ng-container>
-  <ng-template #isLoading>
-    <section fxLayout="column" fxLayoutAlign="center center">
-      <mat-spinner></mat-spinner>
-    </section>
-  </ng-template>
-</article>
