--- libs/import/src/lib/view-extracted-elements/movies/view-extracted-movies.component.scss
@@ -1,14 +1,0 @@
-article {
-    padding: 20px;
-    min-height: calc(100vh - (var(--toolbarHeight) * 2) - 30px); // TODO #1212
-}
-
-.result-tables {
-    h3 {
-        margin-top: 20px;
-    }
-}
-
-mat-spinner {
-    margin: auto;
-}
