+++ libs/import/src/lib/view-extracted-elements/titles/view-extracted-titles.component.html
@@ -1,0 +1,24 @@
+
+<ng-container *ngIf="(moviesToCreate$ | async) || (moviesToUpdate$ | async); else isLoading">
+  <section class="result-tables">
+
+    <ng-container *ngIf="moviesToCreate$ | async as moviesToCreate">
+      <ng-container *ngIf="moviesToCreate.data.length">
+        <h3>Titles to Create</h3>
+        <import-table-extracted-movies [rows]="moviesToCreate" mode="create"></import-table-extracted-movies>
+      </ng-container>
+    </ng-container>
+    <ng-container *ngIf="moviesToUpdate$ | async as moviesToUpdate">
+      <ng-container *ngIf="moviesToUpdate.data.length">
+        <h3>Titles to Update & Submit</h3>
+        <import-table-extracted-movies [rows]="moviesToUpdate" mode="update"></import-table-extracted-movies>
+      </ng-container>
+    </ng-container>
+
+  </section>
+</ng-container>
+<ng-template #isLoading>
+  <section fxLayout="column" fxLayoutAlign="center center">
+    <mat-spinner></mat-spinner>
+  </section>
+</ng-template>
