+++ libs/import/src/lib/view-extracted-elements/titles/view-extracted-titles.component.scss
@@ -1,0 +1,13 @@
+article {
+  padding: 20px;
+}
+
+.result-tables {
+  h3 {
+    margin-top: 20px;
+  }
+}
+
+mat-spinner {
+  margin: auto;
+}
